+++ packages/abi/lib/fragments.js
@@ -510,6 +510,10 @@ function parseModifiers(value, params) {
                 params.payable = true;
                 params.stateMutability = "payable";
                 break;
+            case "nonpayable":
+                params.payable = false;
+                params.stateMutability = "nonpayable";
+                break;
             case "pure":
                 params.constant = true;
                 params.stateMutability = "pure";